chore(src): reorganize Textarea elements
